Create FinalProjectII.pde
This is my latest version. Still working on timings...It may be something real simple but I ran out of time for the day...I will work some more on it tomorrow...I hope this is enough, though, to enable you to get started if you are working on this project this week.